Use `performance.now()`, set groundwork for JS self-profiling
