+++ extensions/ui-profiler/.prettierignore
@@ -4,4 +4,4 @@ node_modules
 **/package.json
 **/.ipynb_checkpoints/**
 **/*.profile.json
-jupyterlab-ui-profiler
+jupyterlab_ui_profiler
