+++ extensions/ui-profiler/jupyter-config/nb-config/jupyterlab_ui_profiler.json
@@ -1,0 +1,7 @@
+{
+  "NotebookApp": {
+    "nbserver_extensions": {
+      "jupyterlab_ui_profiler": true
+    }
+  }
+}
