+++ extensions/ui-profiler/jupyter-config/server-config/jupyterlab_ui_profiler.json
@@ -1,0 +1,7 @@
+{
+  "ServerApp": {
+    "jpserver_extensions": {
+      "jupyterlab_ui_profiler": true
+    }
+  }
+}
